% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Linear_Regression.R
\name{no_info}
\alias{no_info}
\title{Bayesian Linear Regression with  noninformative prior.}
\usage{
no_info(formula, data)
}
\arguments{
\item{formula}{A string fromula.}

\item{data}{A data frame with your data including response.}
}
\description{
Bayesian Linear Regression with  noninformative prior.
}
\examples{
test <- no_info(formula = 'dist ~ speed', data = cars)
test$classic_coef
test$classic_sigma
test$bayesian_coef
test$bayesian_sigma
}
\references{
L. Fahrmeir, T. Kneib, S. Lang, and B. Marx, Regression: Models, Methods and Applications. Springer-Verlag GmbH, 2013.
}
\author{
Camilo Gonzalez <cigonzalez2@uc.cl>
}
