% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Linear_Regression.R
\name{posterior_nig}
\alias{posterior_nig}
\title{Bayesian Linear Regression with Normal-Inverse Gamma prior.}
\usage{
posterior_nig(formula, data, m, M, a, b)
}
\arguments{
\item{formula}{A string fromula.}

\item{data}{A data frame with your data including response.}

\item{m}{A matrix with the hyperparameter of the mean for the normal distribution.}

\item{M}{A matrix with the hyperparameter of the covariance for the normal distribution.}

\item{a}{A number of hyperparameter for the gamma distribution.}

\item{b}{A number of hyperparameter for the gamma distribution.}
}
\value{
A list with the estimated coefficients and standard deviations of the model from the Classical and Bayesian approach
}
\description{
Bayesian Linear Regression with Normal-Inverse Gamma prior.
}
\examples{
# Noninformative Prior (same results Bayesian and Classical Linear Regression)
m <- matrix(0, nrow = 2, ncol = 1)
M <- diag(10000000000000000000, 2)
a <- 0.000000000000000000000000001
b <- 0.000000000000000000000000001
test <- posterior_nig(
formula = 'dist ~ speed', 
data = cars,
m = m,
M = M,
a = a,
b = b 
)
test$classic_coef
test$classic_sigma
test$bayesian_coef
test$bayesian_sigma
}
