% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayesian_Linear_Regression.R
\name{zellner}
\alias{zellner}
\title{Bayesian Linear Regression with  Zellner's prior.}
\usage{
zellner(formula, data, m1 = m1, g = g, a = a, b = b)
}
\arguments{
\item{formula}{A string fromula.}

\item{data}{A data frame with your data including response.}

\item{m1}{A number of hyperparameter of the intercept.}

\item{g}{A string of hyperparameter of the Zellner's prior. "typeI" is 1/n, "typeII" is 1/k^2 and "typeIII" is 1/max(n,k^2).}

\item{a}{A number of hyperparameter for the gamma distribution.}

\item{b}{A number of hyperparameter for the gamma distribution.}
}
\value{
A list with the estimated coefficients and standard deviations of the model from the Classical and Bayesian approach
}
\description{
Bayesian Linear Regression with  Zellner's prior.
}
\examples{
a <- 0.0001
b <- 0.0001
test <- zellner(
formula = 'dist ~ speed', 
data = cars,
m1 = 0,
g = 'typeI',
a = a,
b = b 
)
test$classic_coef
test$classic_sigma
test$bayesian_coef
test$bayesian_sigma
}
\references{
L. Fahrmeir, T. Kneib, S. Lang, and B. Marx, Regression: Models, Methods and Applications. Springer-Verlag GmbH, 2013.
}
\author{
Camilo Gonzalez <cigonzalez2@uc.cl>
}
